% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author_name.R
\name{add_author_name}
\alias{add_author_name}
\title{add_author_name}
\usage{
add_author_name(family = "Family Name", given = "Firstname)
}
\arguments{
\item{character}{Input names as character}
}
\description{
Add author names to highlight in the publication list
}
\examples{
  \dontrun{
  add_author_name(family = "Family Name", given = "Firstname")}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
