% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publistR_knit.R
\name{publistR_knit}
\alias{publistR_knit}
\title{publistR_knit}
\usage{
publistR_knit()
}
\description{
Add author names to highlight in the publication list
}
\examples{
  \dontrun{
  publistR_knit()}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
