% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publistR_section.R
\name{publistR_section}
\alias{publistR_section}
\title{publistR_section}
\usage{
publistR_section(title = "Section title", DOIs = c("vector of DOIs"))
}
\arguments{
\item{title}{The title of the section}

\item{DOIs}{A DOI or list of DOIs for the papers in the section}
}
\description{
Create section for the publication list
}
\examples{
  \dontrun{
  publistR_section(title = "First authorship", DOIs = "https://doi.org/10.1016/j.jacc.2019.06.057")}
}
\author{
Zheer Kejlberg Al-Mashhadi
}
